change chart routes

+++ routes/chart/chart.py
@@ -10,13 +10,19 @@ chart_bp = Blueprint('chart', __name__)
 def save_chart():
     try:
         data = request.json
-        print('data: ', data)
         coin_bot_id = data.get('coin_bot_id')
+        pair = data.get('pair')
+        temporality = data.get('temporality')
 
-        if not coin_bot_id:
-            return jsonify({'success': False, 'message': 'Coin ID required'}), 400
+        if not coin_bot_id or not pair or not temporality:
+            return jsonify({'success': False, 'message': 'One or more fields are missing'}), 400
+        
+        token = data.get('token')
+        pair=data.get('pair')
 
-        print('coin_bot_id: ', coin_bot_id)
+        if token.casefold() == 'btc' and pair == 'btc':
+            return jsonify({'success': False, 'message': 'Coin/pair not valid'}), 400
+        
         delete_last_chart = session.query(Chart).filter_by(coin_bot_id=coin_bot_id).delete()
         print(f"{delete_last_chart} rows deleted")
 
@@ -29,6 +35,9 @@ def save_chart():
             resistance_2=data.get('resistance_2'),
             resistance_3=data.get('resistance_3'),
             resistance_4=data.get('resistance_4'),
+            token=token,
+            pair=pair,
+            temporality=data.get('temporality'),
             coin_bot_id=coin_bot_id
         )
 
@@ -62,7 +71,10 @@ def get_chart_values_by_coin_bot_id(coin_bot_name):
                 'resistance_1': chart.resistance_1,
                 'resistance_2': chart.resistance_2,
                 'resistance_3': chart.resistance_3,
-                'resistance_4': chart.resistance_4
+                'resistance_4': chart.resistance_4,
+                'token': chart.token,
+                'pair': chart.pair,
+                'temporality': chart.temporality
             }
 
             return jsonify({'success': True, 'chart_values': chart_values})
@@ -78,13 +90,18 @@ def get_chart_values_by_coin_bot_id(coin_bot_name):
 # ----- ROUTE FOR THE DASHBOARD ---------------------------
 # Gets the support and resistance lines of a requested coin
 # this route is duplicated with the previous one, just for convenience, as the dashboard is passing the ID of the coin.
-@chart_bp.route('/api/coin-support-resistance/dashboard/<coin_bot_name>', methods=['GET'])
-def get_s_and_r(coin_bot_name):
+@chart_bp.route('/api/coin-support-resistance/dashboard', methods=['GET'])
+def get_s_and_r():
 
     try:
-           
-        coinbot = session.query(CoinBot).filter(CoinBot.bot_id==coin_bot_name).first()
-        chart = session.query(Chart).filter_by(coin_bot_id=coinbot.bot_id).first()
+        coin_id=request.args.get('coin_id')
+        temporality=request.args.get('temporality')
+        pair=request.args.get('pair')
+
+        if coin_id is None or temporality is None or pair is None:
+            return jsonify({'success': False, 'message': 'One or more values are missing'}), 400
+
+        chart = session.query(Chart).filter(Chart.coin_bot_id==coin_id,Chart.pair==pair.casefold(),Chart.temporality==temporality.casefold()).first()
 
         if chart:
             chart_values = {
@@ -95,16 +112,19 @@ def get_s_and_r(coin_bot_name):
                 'resistance_1': chart.resistance_1,
                 'resistance_2': chart.resistance_2,
                 'resistance_3': chart.resistance_3,
-                'resistance_4': chart.resistance_4
+                'resistance_4': chart.resistance_4,
+                'token': chart.token,
+                'pair': chart.pair,
+                'temporality': chart.temporality
             }
 
-            return jsonify({'success': True, 'chart_values': chart_values})
+            return jsonify({'success': True, 'chart_values': chart_values}), 200
         else:
-            return jsonify({'success': False, 'message': 'Chart not found for the given coin ID'})
+            return jsonify({'success': False, 'message': 'Chart not found for the given coin ID'}), 204
 
     except Exception as e:
         session.rollback()
-        return jsonify({'success': False, 'message': str(e)})
+        return jsonify({'success': False, 'message': str(e)}), 500
